Restructure VectorBench scenario
