+++ src/main/java/org/polypheny/simpleclient/scenario/vectorbench/queryBuilder/dml/InsertMetadata.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package org.polypheny.simpleclient.scenario.vectorbench.queryBuilder.insertion;
+package org.polypheny.simpleclient.scenario.vectorbench.queryBuilder.dml;
 
 import com.google.gson.JsonObject;
 import java.util.HashMap;
