+++ src/main/java/org/polypheny/simpleclient/scenario/vectorbench/queryBuilder/dql/SimpleKnnBooleanFeature.java
@@ -1,7 +1,7 @@
 /*
  * The MIT License (MIT)
  *
- * Copyright (c) 2019-4/17/26, 2:54 PM The Polypheny Project
+ * Copyright (c) 2019-5/26/26, 5:15 PM The Polypheny Project
  *
  * Permission is hereby granted, free of charge, to any person obtaining a
  * copy of this software and associated documentation files (the "Software"), to deal
@@ -22,107 +22,90 @@
  * SOFTWARE.
  */
 
-package org.polypheny.simpleclient.scenario.vectorbench.queryBuilder;
+package org.polypheny.simpleclient.scenario.vectorbench.queryBuilder.dql;
 
 import kong.unirest.core.HttpRequest;
 import org.apache.commons.lang3.tuple.ImmutablePair;
 import org.polypheny.simpleclient.query.Query;
 import org.polypheny.simpleclient.query.QueryBuilder;
-import java.util.HashMap;
+import java.util.Arrays;
 import java.util.Map;
 import java.util.Random;
 
-public class MetadataKnnRealCrossJoin extends QueryBuilder {
+public class SimpleKnnBooleanFeature extends QueryBuilder {
 
     private static final boolean EXPECT_RESULT = true;
-
     private final int dimension;
     private final int limit;
-    private final String norm;
-
+    private final String norm; // Should be 'JACCARD' or 'HAMMING'
     private final Random random;
 
-
-    public MetadataKnnRealCrossJoin( long randomSeed, int dimension, int limit, String norm ) {
+    public SimpleKnnBooleanFeature( long randomSeed, int dimension, int limit, String norm ) {
         this.dimension = dimension;
         this.limit = limit;
         this.norm = norm;
-
         this.random = new Random( randomSeed );
     }
 
-
-    private Float[] getRandomVector() {
-        Float[] floats = new Float[this.dimension];
+    private Boolean[] getRandomVector() {
+        Boolean[] booleans = new Boolean[this.dimension];
         for ( int i = 0; i < this.dimension; i++ ) {
-            floats[i] = random.nextInt( 100 ) / 100.0f;
+            booleans[i] = random.nextBoolean();
         }
-
-        return floats;
+        return booleans;
     }
-
 
     @Override
     public synchronized Query getNewQuery() {
-        return new MetadataKnnRealCrossJoin.MetadataKnnRealCrossJoinQuery(
-                getRandomVector(),
-                limit,
-                norm
-        );
+        return new SimpleKnnBooleanFeatureQuery( getRandomVector(), limit, norm );
     }
 
+    private static class SimpleKnnBooleanFeatureQuery extends Query {
 
-    private static class MetadataKnnRealCrossJoinQuery extends Query {
-
-        private static final String SQL_1 = "SELECT knn_metadata.id, knn_metadata.textdata, closest.dist FROM knn_metadata, ( SELECT t1.id, distance(t1.feature, t2.feature, '";
-        private static final String SQL_2 = "') AS dist FROM knn_realfeature t1, knn_realfeature t2 WHERE t2.id = 1 ORDER BY dist ASC LIMIT ";
-        private static final String SQL_3 = ") AS closest WHERE knn_metadata.id = closest.id ORDER BY closest.dist ASC";
-        private final Float[] target;
+        private final Boolean[] target;
         private final int limit;
         private final String norm;
 
-
-        private MetadataKnnRealCrossJoinQuery( Float[] target, int limit, String norm ) {
+        public SimpleKnnBooleanFeatureQuery( Boolean[] target, int limit, String norm ) {
             super( EXPECT_RESULT );
             this.target = target;
             this.limit = limit;
             this.norm = norm;
         }
 
-
         @Override
         public String getSql() {
-            return SQL_1 + norm + SQL_2 + limit + SQL_3;
+            return "SELECT id, " + norm.toLowerCase() + "_distance(feature, ARRAY" + Arrays.toString( target ) + ") "
+                    + "as dist " +
+                    "FROM knn_booleanfeature " + // Ensure this tableexists in DataGenerator
+                    "ORDER BY dist ASC LIMIT " + limit;
         }
 
 
         @Override
-        public String getParameterizedSqlQuery() {
-            return null;
-            //return SQL_1 + "?" + SQL_2 + "'" + norm + "'" + SQL_3 + limit + SQL_4;
-        }
+        public String getParameterizedSqlQuery() { return null; }
 
 
         @Override
-        public Map<Integer, ImmutablePair<DataTypes, Object>> getParameterValues() {
-            Map<Integer, ImmutablePair<DataTypes, Object>> map = new HashMap<>();
-            map.put( 1, new ImmutablePair<>( DataTypes.ARRAY_REAL, target ) );
-            return map;
-        }
+        public Map<Integer, ImmutablePair<DataTypes, Object>>
+        getParameterValues() { return null; }
 
 
         @Override
-        public HttpRequest<?> getRest() {
-            return null;
-        }
+        public HttpRequest<?> getRest() { return null; }
 
 
         @Override
         public String getMongoQl() {
-            return null;
+            return "db.knn_booleanfeature.aggregate([{" +
+                    "    \"$vectorSearch\": {" +
+                    "        \"path\": \"feature\"," +
+                    "        \"queryVector\": " + Arrays.toString( target )
+                    + "," +
+                    "        \"metric\": \"" + norm + "\"," +
+                    "        \"limit\": " + limit +
+                    "    }" +
+                    "}])";
         }
-
     }
-
 }
-
